+++ python/mujincontrollerclient/__init__.py
@@ -127,7 +127,16 @@ class AuthenticationError(ClientExceptionBase):
 
 
 class ControllerClientError(ClientExceptionBase):
-    pass
+
+    _response = None # http response that resulted in the error
+
+    def __init__(self, message='', response=None):
+        super(ControllerClientError, self).__init__(message)
+        self._response = response
+
+    @property
+    def response(self):
+        return self._response
 
 
 class URIError(ClientExceptionBase):
@@ -140,11 +149,13 @@ class ControllerGraphClientException(ClientExceptionBase):
 
     _statusCode = None
     _content = None
+    _response = None
 
-    def __init__(self, message='', statusCode=None, content=None):
+    def __init__(self, message='', statusCode=None, content=None, response=None):
         super(ControllerGraphClientException, self).__init__(message)
         self._statusCode = statusCode
         self._content = content
+        self._response = response
 
     @property
     def statusCode(self):
@@ -153,3 +164,7 @@ class ControllerGraphClientException(ClientExceptionBase):
     @property
     def content(self):
         return self._content
+
+    @property
+    def response(self):
+        return self._response