+++ CHANGELOG.md
@@ -1,10 +1,26 @@
+# 0.16.5 (2022-11-01)
+
+- Add `UploadFiles` to support mult-file upload.
+
+
+# 0.16.3 (2022-11-07)
+
+- Regenerate graph client.
+
+
+# 0.16.2 (2022-11-07)
+
+- Regenerate graph client.
+
 # 0.16.1 (2022-10-07)
 
 - Remove `ExecuteTrajectory` method, to avoid dangerous usage. Instead, an ITL program (e.g. with `Move`) should be executed to ensure that the trajectory is collision-free.
 
+
 # 0.16 (2022-10-06)
 
 - Remove `GetInertiaChildJointStartValues`
+
 
 # 0.15.9 (2022-09-30)